orton/orton/requirements.txt
django>=1.4.2
django-icybackup
django-cms>=2.2
django-cms-redirects
sorl-thumbnail
django-haystack
whoosh
django-cms-search
django-dbsettings
django-reversion==1.6.1
django-tinymce
pil>=1.1.7
easy_thumbnails
django-filer
cmsplugin-filer
cmsplugin-yandexmap
django-dbsettings
django-smart-load-tag
mysql-python